improved fer code; no functionality changes
