+++ facial-expression-rec/fer/test/test_image_copyright_note.txt
@@ -1,0 +1,5 @@
+test_image.png was found with CC0 license
+(https://creativecommons.org/publicdomain/zero/1.0/)
+on
+https://freerangestock.com/photos/57266/photo-details.html,
+https://freerangestock.com/sample/56854/photo.jpg).
