+++ facial-expression-rec/requirements.txt
@@ -5,4 +5,3 @@ websockets~=10.4
 numpy~=1.25.2
 onnxruntime~=1.15.1
 opencv-python~=4.8.0.76
-Pillow~=10.0.0
